+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Unreleased Changes
+------------------
+
+* Minor fix to instructions printed after release build in ``biweeklybudget/tests/docker_build.py``
+
 0.1.1 (2017-05-20)
 ------------------